+++ internal/services/compute/managed_disk_resource_test.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/Azure/azure-sdk-for-go/profiles/2019-03-01/compute/mgmt/compute"
+	"github.com/Azure/azure-sdk-for-go/profiles/2020-09-01/compute/mgmt/compute"
 	"github.com/hashicorp/go-azure-helpers/lang/pointer"
 	"github.com/hashicorp/terraform-provider-azurestack/internal/clients"
 	"github.com/hashicorp/terraform-provider-azurestack/internal/services/compute/parse"
@@ -249,7 +249,8 @@ func (ManagedDiskResource) destroyVirtualMachine(ctx context.Context, client *cl
 	vmName := state.Attributes["name"]
 	resourceGroup := state.Attributes["resource_group_name"]
 
-	future, err := client.Compute.VMClient.Delete(ctx, resourceGroup, vmName)
+	var forceDeletion *bool = nil
+	future, err := client.Compute.VMClient.Delete(ctx, resourceGroup, vmName, forceDeletion)
 	if err != nil {
 		return fmt.Errorf("Bad: Delete on vmClient: %+v", err)
 	}
